+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 2.0.0beta - TBD
+- `allowFootnote` becomes `allowReference`
+- `onFootnote` becomes `onReference`
+
 ## 1.0.0 - 2020-08-13
-* Initial version
+- Initial version